+++ programs/spl-token-bonding/src/lib.rs
@@ -205,7 +205,7 @@ pub mod spl_token_bonding {
       ))?;
     }
 
-    msg!("Setting mint authority to none");
+    msg!("Setting mint authority to general authority");
     if ctx.accounts.target_mint.mint_authority.is_some()
       && ctx.accounts.target_mint.mint_authority.unwrap() == token_bonding.key()
     {
@@ -219,7 +219,7 @@ pub mod spl_token_bonding {
           bonding_seeds,
         ),
         spl_token::instruction::AuthorityType::MintTokens,
-        None,
+        Some(ctx.accounts.general_authority.key()),
       )?;
     }
 
@@ -251,9 +251,20 @@ pub mod spl_token_bonding {
     let base_mint = &ctx.accounts.base_mint;
     let target_mint = &ctx.accounts.target_mint;
     let curve = &ctx.accounts.curve;
-    let base_amount = precise_supply_amt(ctx.accounts.base_storage.amount, base_mint);
+    // In the case of a sell frozen bonding curve, we don't want to be adaptive relative to the base storage
+    // amount.
+    let base_amount_u64 = if token_bonding.sell_frozen {
+      token_bonding.reserve_balance_from_bonding
+    } else {
+      ctx.accounts.base_storage.amount
+    };
+    let base_amount = precise_supply_amt(base_amount_u64, base_mint);
     let target_supply = precise_supply(target_mint);
 
+    // Not yet initialized since reserve_balance_from_bonding is a new feature
+    if ctx.accounts.target_mint.supply > 0 && token_bonding.reserve_balance_from_bonding == 0 {
+      token_bonding.reserve_balance_from_bonding = ctx.accounts.base_storage.amount;
+    }
     msg!(
       "Current reserves {} and supply {}",
       ctx.accounts.base_storage.amount,
@@ -424,6 +435,7 @@ pub mod spl_token_bonding {
     }
 
     msg!("Paying out {} to base storage", price);
+    token_bonding.reserve_balance_from_bonding = token_bonding.reserve_balance_from_bonding + price;
     token::transfer(
       CpiContext::new(
         token_program.clone(),
@@ -598,6 +610,7 @@ pub mod spl_token_bonding {
       reclaimed,
       ctx.accounts.base_storage.amount
     );
+    token_bonding.reserve_balance_from_bonding = token_bonding.reserve_balance_from_bonding - reclaimed;
     token::transfer(
       CpiContext::new_with_signer(
         token_program.clone(),